added navbar component for desktop view
